% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_lppls.R
\name{apply_ds_filters}
\alias{apply_ds_filters}
\title{Apply DS LPPLS Filtering Criteria (Table 3.1 from thesis)}
\usage{
apply_ds_filters(scale_results, bubble_type, filter_type)
}
\description{
Apply DS LPPLS Filtering Criteria (Table 3.1 from thesis)
}
\keyword{internal}
