% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_lppls.R
\name{calculate_ds_confidence_indicators}
\alias{calculate_ds_confidence_indicators}
\title{Calculate DS LPPLS Confidence Indicators Following Original Methodology}
\usage{
calculate_ds_confidence_indicators(window_results, current_date)
}
\description{
Calculate DS LPPLS Confidence Indicators Following Original Methodology
}
\keyword{internal}
