% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_estimation.R
\name{calculate_linear_parameters}
\alias{calculate_linear_parameters}
\title{Calculate Linear Parameters for LPPLS Model}
\usage{
calculate_linear_parameters(data, tc, m, w)
}
\arguments{
\item{data}{Data frame with time series data}

\item{tc}{Critical time parameter}

\item{m}{Power law exponent}

\item{w}{Log-periodic frequency}
}
\value{
Vector of linear parameters \link{A, B, C1, C2}
}
\description{
Calculate Linear Parameters for LPPLS Model
}
