% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_utils.R
\name{create_timer}
\alias{create_timer}
\title{Create Benchmark Timer}
\usage{
create_timer(description = "LPPLS Analysis")
}
\arguments{
\item{description}{Description of the benchmark}
}
\value{
List with timer functions
}
\description{
Create Benchmark Timer
}
