% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{generate_synthetic_data}
\alias{generate_synthetic_data}
\title{Generate Synthetic Financial Data}
\usage{
generate_synthetic_data(type = "sp500", n = 1000, start_date = "2020-01-01")
}
\arguments{
\item{type}{Type of data to generate}

\item{n}{Number of observations (default: 1000)}

\item{start_date}{Start date for the series (default: "2020-01-01")}
}
\value{
Data frame with synthetic financial data
}
\description{
Generate Synthetic Financial Data
}
