% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_lppls.R
\name{initialize_ds_confidence_structure}
\alias{initialize_ds_confidence_structure}
\title{Initialize DS Confidence Structure for a Single Day}
\usage{
initialize_ds_confidence_structure(current_date)
}
\description{
Initialize DS Confidence Structure for a Single Day
}
\keyword{internal}
