% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load Sample Data}
\usage{
load_sample_data(dataset = "sp500")
}
\arguments{
\item{dataset}{Name of sample dataset to load}
}
\value{
Data frame with sample financial data
}
\description{
Load Sample Data
}
