% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_estimation.R
\name{lppls_de_optimize}
\alias{lppls_de_optimize}
\title{Differential Evolution Optimization for LPPLS}
\usage{
lppls_de_optimize(start_params, lower, upper, data)
}
\arguments{
\item{start_params}{Initial parameter vector \link{tc, m, w}}

\item{lower}{Lower bounds vector}

\item{upper}{Upper bounds vector}

\item{data}{Time series data frame}
}
\value{
Optimization result compatible with MLSL output format
}
\description{
Differential Evolution Optimization for LPPLS
}
