% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_fit.R
\name{lppls_fit}
\alias{lppls_fit}
\title{Estimates the LPPLS parameters (single time window)}
\usage{
lppls_fit(data)
}
\arguments{
\item{data}{Time series dataset: First column (class Date) in format "\%Y-\%m-\%d" and second column is the asset price}

\item{t1}{Beginning of time windows upon which estimate the parameters}

\item{t2}{End of time windows upon which estimate the parameters (If t2 = "today" the last day in the time series will be used)}

\item{type}{Optimization methods ("L-BFGS-B", "CMAES", "contr", "isres", "mlsl", "nelder")}

\item{plot}{Generate a plot if TRUE}
}
\value{
List containing parameters and other data
}
\description{
Estimates the LPPLS parameters (single time window)
}
\examples{
lppls_fit(data)
}
