% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_estimation.R
\name{lppls_model}
\alias{lppls_model}
\title{LPPLS Model Function}
\usage{
lppls_model(data, tc, m, w, a, b, c1, c2)
}
\arguments{
\item{data}{Data frame with time series data}

\item{tc}{Critical time parameter}

\item{m}{Power law exponent}

\item{w}{Log-periodic frequency}

\item{a}{Constant term}

\item{b}{Linear coefficient}

\item{c1}{Cosine coefficient}

\item{c2}{Sine coefficient}
}
\value{
Vector of fitted values
}
\description{
LPPLS Model Function
}
