% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_estimation.R
\name{lppls_objective_function}
\alias{lppls_objective_function}
\title{LPPLS Objective Function}
\usage{
lppls_objective_function(x, data)
}
\arguments{
\item{x}{Vector of parameters \link{tc, m, w}}

\item{data}{Data frame with time series data}
}
\value{
Residual sum of squares
}
\description{
LPPLS Objective Function
}
