% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_lppls_base}
\alias{plot_lppls_base}
\title{Internal function for base R LPPLS fit plotting}
\usage{
plot_lppls_base(data, lppls_result, title, show_critical_time)
}
\description{
Internal function for base R LPPLS fit plotting
}
\keyword{internal}
