% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_lppls_fit}
\alias{plot_lppls_fit}
\title{Plot LPPLS Model Fit}
\usage{
plot_lppls_fit(
  data,
  lppls_result,
  title = "LPPLS Model Fit",
  show_critical_time = TRUE
)
}
\arguments{
\item{data}{Original time series data}

\item{lppls_result}{Result from lppls_estimate function}

\item{title}{Plot title (optional)}

\item{show_critical_time}{Logical, whether to show critical time line}
}
\value{
ggplot object if ggplot2 is available, otherwise base plot
}
\description{
Plot LPPLS Model Fit
}
