% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_lppls_ggplot}
\alias{plot_lppls_ggplot}
\title{Internal function for ggplot2-based LPPLS fit plotting}
\usage{
plot_lppls_ggplot(data, lppls_result, title, show_critical_time)
}
\description{
Internal function for ggplot2-based LPPLS fit plotting
}
\keyword{internal}
