% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_multiple_lppls}
\alias{plot_multiple_lppls}
\title{Plot Multiple LPPLS Models}
\usage{
plot_multiple_lppls(
  data,
  models_list,
  title = "Multiple LPPLS Models",
  max_models = 15
)
}
\arguments{
\item{data}{Original time series data}

\item{models_list}{List of LPPLS model results}

\item{title}{Plot title (optional)}

\item{max_models}{Maximum number of models to plot (default: 15)}
}
\value{
Plot showing multiple LPPLS fits
}
\description{
Plot Multiple LPPLS Models
}
