% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{validate_lppls_data}
\alias{validate_lppls_data}
\title{Validate LPPLS Data}
\usage{
validate_lppls_data(data)
}
\arguments{
\item{data}{Data frame to validate}
}
\value{
Logical indicating if data is valid, with attributes for any issues
}
\description{
Validate LPPLS Data
}
