% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppls_estimation.R
\name{lppls_estimate}
\alias{lppls_estimate}
\title{LPPLS Model Estimation Function}
\usage{
lppls_estimate(data, plot = FALSE)
}
\arguments{
\item{data}{A data frame with columns Date, Close, and t (decimal time)}

\item{plot}{Logical, whether to plot the fitted model}
}
\value{
A data frame with estimation results and parameters
}
\description{
LPPLS Model Estimation Function
}
